Replace `request.site` with `Site.find_for_req..`
The `SiteMiddleware` is deprecated as of Wagtail 2.11. The new way to
find the site for a request is `Site.find_for_request(request)`.
This commit switches the the two usages of `request.site` to the new
way.

No other usecases of `.site` on a request object were found by running
`ag .site --py` on the repo.