Update SiteMiddleware path in MIDDLEWARE list
The reliance on the SiteMiddleware in this repo itself has been removed
by switching from `request.site` to `Site.find_for_request(request)`.

But, tests where failing when the `SiteMiddleware` was removed from the
middleware list. This probably means that some other package still
depends on it.

For those packages to work, the path for the `SiteMiddleware` was
updated to the legacy package that is still supported.